Nav active link demo
When a user clicks on a specific link in the navbar, it should get highlighted. For that, we can use "NavLink" tag which automatically applies the active class on the clicked link element. Then you can target the specific element using the css selector with active class on it and style it.